use enums with categories instead of ints to avoid SQLite casting

+++ src/slskd/Transfers/API/Controllers/TransfersController.cs
@@ -117,7 +117,7 @@ namespace slskd.Transfers.API
 
             try
             {
-                Transfers.Downloads.Remove(t => !t.Removed && TransferStateCategories.Completed.Contains((int)t.State));
+                Transfers.Downloads.Remove(t => !t.Removed && TransferStateCategories.Completed.Contains(t.State));
                 return NoContent();
             }
             catch (Exception ex)
@@ -186,7 +186,7 @@ namespace slskd.Transfers.API
 
             try
             {
-                Transfers.Uploads.Remove(t => !t.Removed && TransferStateCategories.Completed.Contains((int)t.State));
+                Transfers.Uploads.Remove(t => !t.Removed && TransferStateCategories.Completed.Contains(t.State));
                 return NoContent();
             }
             catch (Exception ex)
